add data ans services

+++ src/drivers/DHT_driver.cpp
@@ -1,32 +1,32 @@
 #include "DHT_driver.h"
 
 
-int DHT_driver::get_temperature()
+int8_t DHT_driver::get_temperature()
 {
     return dht.readTemperature();
 }
 
-int DHT_driver::get_temperature(int addr)
+int8_t DHT_driver::get_temperature(int addr)
 {
     return get_temperature();
 }
 
-int DHT_driver::get_temperature_middle()
+int8_t DHT_driver::get_temperature_middle()
 {
     return get_temperature();
 }
 
-int DHT_driver::get_temperature_delta()
+int8_t DHT_driver::get_temperature_delta()
 {
     return get_temperature();
 }
 
-int DHT_driver::get_humidity()
+uint8_t DHT_driver::get_humidity()
 {
-    return dht.readHumidity();
+    return uint8_t(dht.readHumidity());
 }
 
-int DHT_driver::get_humidity(int addr)
+uint8_t DHT_driver::get_humidity(int addr)
 {
-    return dht.readHumidity();
+    return uint8_t(dht.readHumidity());
 }
